+++ src/commands/service/add.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { isAbsolute, relative, resolve } from "node:path";
+import { isAbsolute, resolve } from "node:path";
 import { cwd } from "node:process";
 
 import { Args, Flags } from "@oclif/core";
@@ -23,14 +23,16 @@ import { BaseCommand, baseFlags } from "../../baseCommand.js";
 import { addService } from "../../lib/addService.js";
 import { commandObj } from "../../lib/commandObj.js";
 import { initReadonlyServiceConfig } from "../../lib/configs/project/service.js";
-import { FLUENCE_CONFIG_FULL_FILE_NAME } from "../../lib/const.js";
+import {
+  FLUENCE_CONFIG_FULL_FILE_NAME,
+  MARINE_BUILD_ARGS,
+} from "../../lib/const.js";
 import {
   AQUA_NAME_REQUIREMENTS,
   isUrl,
 } from "../../lib/helpers/downloadFile.js";
 import { initCli } from "../../lib/lifeCycle.js";
 import { initMarineCli } from "../../lib/marineCli.js";
-import { projectRootDir } from "../../lib/paths.js";
 import { input } from "../../lib/prompt.js";
 
 const PATH_OR_URL = "PATH | URL";
@@ -44,6 +46,7 @@ export default class Add extends BaseCommand<typeof Add> {
       description: `Override service name (${AQUA_NAME_REQUIREMENTS})`,
       helpValue: "<name>",
     }),
+    ...MARINE_BUILD_ARGS,
   };
   static override args = {
     [PATH_OR_URL]: Args.string({
@@ -67,28 +70,30 @@ export default class Add extends BaseCommand<typeof Add> {
     );
 
     if (serviceConfig === null) {
-      commandObj.error(`No service config at ${serviceOrServiceDirPathOrUrl}`);
+      commandObj.error(
+        `No service config found at ${serviceOrServiceDirPathOrUrl}`,
+      );
     }
 
     const marineCli = await initMarineCli(fluenceConfig);
 
     await addService({
       serviceName: flags.name ?? serviceConfig.name,
-      pathOrUrl: resolveServicePathOrUrl(serviceOrServiceDirPathOrUrl),
+      absolutePathOrUrl: resolveServicePathOrUrl(serviceOrServiceDirPathOrUrl),
       fluenceConfig,
       marineCli,
+      marineBuildArgs: flags["marine-build-args"],
     });
   }
 }
 
 const resolveServicePathOrUrl = (serviceOrServiceDirPathOrUrl: string) => {
-  if (isUrl(serviceOrServiceDirPathOrUrl)) {
+  if (
+    isUrl(serviceOrServiceDirPathOrUrl) ||
+    isAbsolute(serviceOrServiceDirPathOrUrl)
+  ) {
     return serviceOrServiceDirPathOrUrl;
   }
 
-  if (isAbsolute(serviceOrServiceDirPathOrUrl)) {
-    return relative(projectRootDir, serviceOrServiceDirPathOrUrl);
-  }
-
-  return relative(projectRootDir, resolve(serviceOrServiceDirPathOrUrl));
+  return resolve(serviceOrServiceDirPathOrUrl);
 };