+++ src/commands/workers/deploy.ts
@@ -34,6 +34,7 @@ import {
   IMPORT_FLAG,
   NO_BUILD_FLAG,
   TRACING_FLAG,
+  MARINE_BUILD_ARGS,
 } from "../../lib/const.js";
 import {
   ensureAquaFileWithWorkerInfo,
@@ -56,10 +57,11 @@ export default class Deploy extends BaseCommand<typeof Deploy> {
     ...IMPORT_FLAG,
     ...NO_BUILD_FLAG,
     ...TRACING_FLAG,
+    ...MARINE_BUILD_ARGS,
   };
   static override args = {
     "WORKER-NAMES": Args.string({
-      description: `Names of workers to deploy (by default all workers from 'hosts' property in ${FLUENCE_CONFIG_FULL_FILE_NAME} are deployed)`,
+      description: `Comma separated names of workers to deploy. Example: "worker1,worker2" (by default all workers from 'hosts' property in ${FLUENCE_CONFIG_FULL_FILE_NAME} are deployed)`,
     }),
   };
   async run(): Promise<void> {
@@ -83,6 +85,7 @@ export default class Deploy extends BaseCommand<typeof Deploy> {
       maybeWorkersConfig: workersConfig,
       aquaImports,
       noBuild: flags["no-build"],
+      marineBuildArgs: flags["marine-build-args"],
     });
 
     await initFluenceClient(flags, fluenceConfig);