+++ src/lib/configs/project/fluence.ts
@@ -53,8 +53,8 @@ import {
   type Relays,
 } from "../../multiaddres.js";
 import {
-  ensureFluenceDir,
   ensureSrcAquaMainPath,
+  getFluenceDir,
   projectRootDir,
 } from "../../paths.js";
 import { FLUENCE_ENV } from "../../setupEnvironment.js";
@@ -225,6 +225,7 @@ type ConfigV2 = Omit<ConfigV1, "version"> & {
   spells?: Record<string, FluenceConfigSpell>;
   aquaImports?: Array<string>;
   cliVersion?: string;
+  marineBuildArgs?: string;
 };
 
 const spellSchema: JSONSchemaType<FluenceConfigSpell> = {
@@ -261,6 +262,8 @@ const dealSchema: JSONSchemaType<Deal> = {
   required: [],
 } as const;
 
+const validateDealSchema = ajv.compile(dealSchema);
+
 const workerConfigSchema: JSONSchemaType<Worker> = {
   type: "object",
   description: "Worker config",
@@ -292,6 +295,16 @@ const hostConfigSchema: JSONSchemaType<Host> = {
   },
   required: ["peerIds"],
 };
+
+const validateHostsSchema = ajv.compile(hostConfigSchema);
+
+export function assertIsArrayWithHostsOrDeals(
+  unknownArr: [string, unknown][],
+): asserts unknownArr is [string, Host | Deal][] {
+  unknownArr.forEach(([, unknown]) => {
+    assert(validateHostsSchema(unknown) || validateDealSchema(unknown));
+  });
+}
 
 const configSchemaV2: JSONSchemaType<ConfigV2> = {
   ...configSchemaV1Obj,
@@ -419,6 +432,11 @@ const configSchemaV2: JSONSchemaType<ConfigV2> = {
       items: { type: "string" },
       nullable: true,
     },
+    marineBuildArgs: {
+      type: "string",
+      description: `\`cargo build\` arguments to pass to marine build command`,
+      nullable: true,
+    },
     cliVersion: {
       type: "string",
       description: `The version of the ${CLI_NAME_FULL} that is compatible with this project. Set this to enforce a particular set of versions of all fluence components`,
@@ -468,14 +486,16 @@ const getDefaultConfig = async (): Promise<string> => {
 
 # A map with worker names as keys and worker configs as values
 workers:
-  ${DEFAULT_WORKER_NAME}: # worker name
+# # worker name
+  ${DEFAULT_WORKER_NAME}:
     services: [] # list of service names to be deployed to this worker
     spells: [] # list of spell names to be deployed to this worker
 
 
 # A map with worker names as keys and deals as values
 deals:
-  ${DEFAULT_WORKER_NAME}: # worker name
+# # worker name
+  ${DEFAULT_WORKER_NAME}:
     minWorkers: ${MIN_WORKERS} # required amount of workers to activate the deal
     targetWorkers: ${TARGET_WORKERS} # max amount of workers in the deal
 
@@ -497,12 +517,14 @@ version: 2
 # # Service names must start with a lowercase letter and contain only letters numbers and underscores.
 # # You can use \`fluence service new\` or \`fluence service add\` command to add a service
 # services:
-#   myService: # service name
+#   # service name
+#   myService:
 #     # Path to service directory, service config or URL to the tar.gz archive that contains the service
 #     get: "src/services/myService"
 #     # A map of modules that you want to override for this service
 #     overrideModules:
-#       moduleName: # module name
+#       # module name
+#       moduleName:
 #         # environment variables accessible by a particular module
 #         # with standard Rust env API like this: std::env::var(IPFS_ADDR_ENV_NAME)
 #         # Module environment variables could be examined with repl
@@ -539,7 +561,8 @@ version: 2
 #
 # # A map with spell names as keys and spell configs as values
 # spells:
-#   mySpell: # spell name
+#   # spell name
+#   mySpell:
 #     # Path to spell config or directory with spell config
 #     get: "src/spells/mySpell"
 #
@@ -638,7 +661,8 @@ version: 2
 #
 # # if you want to deploy your services to specific peerIds. Soon it will be deprecated in favor of \`deals\` property
 # hosts:
-#   ${DEFAULT_WORKER_NAME}: # worker name
+#   # worker name
+#   ${DEFAULT_WORKER_NAME}:
 #     peerIds:
 #       - ${getDefaultPeerId(DEFAULT_RELAYS_FOR_TEMPLATE)}
 `;
@@ -841,7 +865,7 @@ const initConfigOptions: InitConfigOptions<Config, LatestConfig> = {
   migrations,
   name: FLUENCE_CONFIG_FILE_NAME,
   getConfigOrConfigDirPath,
-  getSchemaDirPath: ensureFluenceDir,
+  getSchemaDirPath: getFluenceDir,
   validate,
 };