+++ src/lib/configs/project/workers.ts
@@ -23,7 +23,7 @@ import {
   WORKERS_CONFIG_FILE_NAME,
   CLI_NAME,
 } from "../../const.js";
-import { ensureFluenceDir } from "../../paths.js";
+import { getFluenceDir } from "../../paths.js";
 import {
   getReadonlyConfigInitFunction,
   getConfigInitFunction,
@@ -166,7 +166,7 @@ const initConfigOptions: InitConfigOptions<Config, LatestConfig> = {
   latestSchema: configSchemaV0,
   migrations,
   name: WORKERS_CONFIG_FILE_NAME,
-  getConfigOrConfigDirPath: ensureFluenceDir,
+  getConfigOrConfigDirPath: getFluenceDir,
 };
 
 const getDefault: GetDefaultConfig = () => {
@@ -178,7 +178,8 @@ version: 0
 
 # # A map of created deals
 # deals:
-#   defaultWorker: # worker name
+#   # worker name
+#   defaultWorker:
 #     # worker CID
 #     definition: bafkreigvy3k4racm6i6vvavtr5mdkllmfi2lfkmdk72gnzwk7zdnhajw4y
 #     # ISO timestamp of the time when the worker was deployed
@@ -194,7 +195,8 @@ version: 0
 
 # # A map of deployed workers
 # hosts:
-#   defaultWorker: # worker name
+#   # worker name
+#   defaultWorker:
 #     # worker CID
 #     definition: bafkreicoctafgctpxf7jk4nynpnma4wdxpcecjtspsjmuidmag6enctnqa
 #     # worker installation spells