+++ src/lib/const.ts
@@ -138,6 +138,8 @@ export const INDEX_JS_FILE_NAME = `index.${JS_EXT}`;
 export const CONFIG_TOML = `Config.${TOML_EXT}`;
 export const CARGO_TOML = `Cargo.${TOML_EXT}`;
 
+export const README_MD_FILE_NAME = `README.md`;
+
 export const FS_OPTIONS = {
   encoding: "utf8",
 } as const;
@@ -233,6 +235,13 @@ export const TRACING_FLAG = {
   }),
 };
 
+export const MARINE_BUILD_ARGS = {
+  "marine-build-args": Flags.string({
+    description: `\`cargo build\` flags and args to pass to marine build. Overrides 'marineBuildArgs' property in ${FLUENCE_CONFIG_FULL_FILE_NAME}`,
+    helpValue: "<--flag arg>",
+  }),
+};
+
 export const TTL_FLAG_NAME = "ttl";
 export const DIAL_TIMEOUT_FLAG_NAME = "dial-timeout";
 
@@ -306,13 +315,14 @@ export const PACKAGE_NAME_AND_VERSION_ARG_NAME =
 
 export const RECOMMENDED_GITIGNORE_CONTENT = `.idea
 .DS_Store
-${DOT_FLUENCE_DIR_NAME}
+/${DOT_FLUENCE_DIR_NAME}/${PROJECT_SECRETS_FULL_CONFIG_FILE_NAME}
+/${DOT_FLUENCE_DIR_NAME}/${SCHEMAS_DIR_NAME}
 **/node_modules
 **/target/
 .repl_history
-.vscode/settings.json
-src/ts/src/aqua
-src/js/src/aqua`;
+/.vscode/settings.json
+/src/ts/src/aqua
+/src/js/src/aqua`;
 
 export const IS_TTY = process.stdout.isTTY && process.stdin.isTTY;
 export const IS_DEVELOPMENT = process.env["NODE_ENV"] === "development";
@@ -366,13 +376,11 @@ func runDeployedServices() -> *Answer:
     dealId = workersInfo.deals.defaultWorker!.dealId
     answers: *Answer
     workers <- resolveSubnetwork(dealId)
-    for w <- workers! par:
+    for w <- workers!:
         on w.metadata.peer_id via w.metadata.relay_id:
             answer <- MyService.greeting("fluence")
             answers <<- Answer(answer=answer, peer=w.metadata.relay_id!)
 
-    join answers[workers!.length - 1]
-    par Peer.timeout(PARTICLE_TTL / 2, "TIMED OUT")
     <- answers`;
 
 const RUN_DEPLOYED_SERVICE_AQUA_COMMENT = aquaComment(
@@ -391,7 +399,7 @@ import "${AQUA_WORKERS_FILE_NAME}"
 import "services.aqua"
 
 -- IMPORTANT: Add exports for all functions that you want to run
-export helloWorld, helloWorldRemote, getInfo, getInfos, getInfosInParallel
+export helloWorld, helloWorldRemote, getInfo, getInfos
 
 -- DOCUMENTATION:
 -- https://fluence.dev
@@ -428,18 +436,6 @@ func getInfos(peers: []PeerId) -> []Info:
         on p:
             infos <- Peer.identify()
     <- infos
-
--- parallel computation
-func getInfosInParallel(peers: []PeerId) -> []Info:
-    infos: *Info
-    for p <- peers par:
-        on p:
-            infos <- Peer.identify()
-
-    join infos[Op.array_length(peers) - 1] -- "-1" because it's 0-based
-    par Peer.timeout(PARTICLE_TTL / 2, "")
-
-    <- infos
 `;
 };
 
@@ -471,7 +467,6 @@ import {
   helloWorldRemote,
   getInfo,
   getInfos,
-  getInfosInParallel,
 } from "./aqua/main.js";
 
 const peerIds = ${NODES_CONST}.map(({ peerId }) => peerId);
@@ -486,7 +481,6 @@ const main = async () => {
   const helloWorldResult = await helloWorld("Fluence");
   const helloWorldRemoteResult = await helloWorldRemote("Fluence");
   const getInfoResult = await getInfo();
-  const getInfosInParallelResult = await getInfosInParallel(peerIds);
 
   console.log(helloWorldResult);
 
@@ -506,3 +500,105 @@ func spell():
     Spell "worker-spell"
     Spell.list_push_string("logs", str)
 `;
+
+const QUICKSTART_README = `# Fluence Quickstart Template
+
+## Usage
+
+\`\`\`sh
+# You can deploy right away with an example worker that contains an example service
+fluence deal deploy
+
+# Run the deployed code
+fluence run -f 'runDeployedServices()'
+\`\`\`
+`;
+
+const MINIMAL_README = `# Fluence Minimal Template
+
+## Usage
+
+\`\`\`sh
+# Generate a service template and add it to the default worker
+fluence service new myService
+
+# Deploy the default worker
+fluence deal deploy
+
+# Uncomment \`runDeployedServices\` aqua function in \`src/aqua/main.aqua\` and run it
+fluence run -f 'runDeployedServices()'
+\`\`\`
+`;
+
+const TS_README = `# Fluence TypeScript Template
+
+## Usage
+
+\`\`\`sh
+# Compile example aqua code to TypeScript
+fluence aqua
+
+# \`cd\` into \`ts\` directory
+cd src/ts
+
+# Install dependencies
+npm i
+
+# Run example code
+npm start
+
+# You can also deploy deal and run the deployed code
+
+# Generate a service template and add it to the default worker
+fluence service new myService
+
+# Deploy the default worker
+fluence deal deploy
+
+# Uncomment \`runDeployedServices\` aqua function in \`src/aqua/main.aqua\` and compile it
+fluence aqua
+
+# Import \`runDeployedServices\` function in \`src/ts/src/index.ts\` and run it
+npm start
+\`\`\`
+`;
+
+const JS_README = `# Fluence JavaScript Template
+
+## Usage
+
+\`\`\`sh
+# Compile example aqua code to JavaScript
+fluence aqua
+
+# \`cd\` into \`js\` directory
+cd src/js
+
+# Install dependencies
+npm i
+
+# Run example code
+npm start
+
+# You can also deploy deal and run the deployed code
+
+# Generate a service template and add it to the default worker
+fluence service new myService
+
+# Deploy the default worker
+fluence deal deploy
+
+# Uncomment \`runDeployedServices\` aqua function in \`src/aqua/main.aqua\` and compile it
+fluence aqua
+
+# Import \`runDeployedServices\` function in \`src/ts/src/index.js\` and run it
+npm start
+\`\`\`
+`;
+
+export const READMEs: Record<Template, string> = {
+  quickstart: QUICKSTART_README,
+  minimal: MINIMAL_README,
+  ts: TS_README,
+  js: JS_README,
+};