+++ src/lib/init.ts
@@ -44,6 +44,7 @@ import {
   FLUENCE_NETWORK_ENVIRONMENT_NPM_DEPENDENCY,
   CLI_NAME_FULL,
   getMainAquaFileContent,
+  READMEs,
 } from "../lib/const.js";
 import { replaceHomeDir } from "../lib/helpers/replaceHomeDir.js";
 import {
@@ -57,8 +58,9 @@ import {
   getGitignorePath,
   projectRootDir,
   setProjectRootDir,
+  getREADMEPath,
 } from "../lib/paths.js";
-import { input, list } from "../lib/prompt.js";
+import { confirm, input, list } from "../lib/prompt.js";
 import versions from "../versions.json" assert { type: "json" };
 
 import { addService } from "./addService.js";
@@ -124,12 +126,8 @@ export const init = async (options: InitArg = {}): Promise<FluenceConfig> => {
             })),
         );
 
-  if (
-    existsSync(projectPath) &&
-    (await stat(projectPath)).isDirectory() &&
-    (await readdir(projectPath)).length > 0
-  ) {
-    return commandObj.error(
+  if (!(await shouldInit(projectPath))) {
+    commandObj.error(
       `Directory ${color.yellow(
         projectPath,
       )} is not empty. Please, init in an empty directory.`,
@@ -142,49 +140,6 @@ export const init = async (options: InitArg = {}): Promise<FluenceConfig> => {
   setProjectRootDir(projectPath);
   await writeFile(await ensureFluenceAquaServicesPath(), "", FS_OPTIONS);
   const fluenceConfig = await initNewFluenceConfig();
-
-  switch (template) {
-    case "quickstart": {
-      const serviceName = "myService";
-      const servicePath = join(await ensureSrcServicesDir(), serviceName);
-      const pathToModuleDir = join(servicePath, "modules", serviceName);
-      await generateNewModule(pathToModuleDir);
-
-      await initNewReadonlyServiceConfig(
-        servicePath,
-        relative(servicePath, pathToModuleDir),
-        serviceName,
-      );
-
-      await addService({
-        serviceName,
-        fluenceConfig,
-        marineCli: await initMarineCli(fluenceConfig),
-        pathOrUrl: relative(projectRootDir, servicePath),
-        interactive: false,
-      });
-
-      break;
-    }
-
-    case "minimal":
-      break;
-
-    case "js": {
-      await initTSorJSProject({ isJS: true, fluenceConfig });
-      break;
-    }
-
-    case "ts": {
-      await initTSorJSProject({ isJS: false, fluenceConfig });
-      break;
-    }
-
-    default: {
-      const _exhaustiveCheck: never = template;
-      return _exhaustiveCheck;
-    }
-  }
 
   await writeFile(
     await ensureSrcAquaMainPath(),
@@ -213,6 +168,55 @@ export const init = async (options: InitArg = {}): Promise<FluenceConfig> => {
 
   const workersConfig = await initNewWorkersConfig();
   await ensureAquaFileWithWorkerInfo(workersConfig, fluenceConfig);
+  await writeFile(getREADMEPath(), READMEs[template], FS_OPTIONS);
+
+  switch (template) {
+    case "quickstart": {
+      const serviceName = "myService";
+
+      const absoluteServicePath = join(
+        await ensureSrcServicesDir(),
+        serviceName,
+      );
+
+      const pathToModuleDir = join(absoluteServicePath, "modules", serviceName);
+      await generateNewModule(pathToModuleDir);
+
+      await initNewReadonlyServiceConfig(
+        absoluteServicePath,
+        relative(absoluteServicePath, pathToModuleDir),
+        serviceName,
+      );
+
+      await addService({
+        serviceName,
+        fluenceConfig,
+        marineCli: await initMarineCli(fluenceConfig),
+        absolutePathOrUrl: absoluteServicePath,
+        interactive: false,
+      });
+
+      break;
+    }
+
+    case "minimal":
+      break;
+
+    case "js": {
+      await initTSorJSProject({ isJS: true, fluenceConfig });
+      break;
+    }
+
+    case "ts": {
+      await initTSorJSProject({ isJS: false, fluenceConfig });
+      break;
+    }
+
+    default: {
+      const _exhaustiveCheck: never = template;
+      return _exhaustiveCheck;
+    }
+  }
 
   commandObj.logToStderr(
     color.magentaBright(
@@ -223,6 +227,48 @@ export const init = async (options: InitArg = {}): Promise<FluenceConfig> => {
   );
 
   return fluenceConfig;
+};
+
+const shouldInit = async (projectPath: string): Promise<boolean> => {
+  if (!isInteractive) {
+    return true;
+  }
+
+  const pathDoesNotExists = !existsSync(projectPath);
+
+  if (pathDoesNotExists) {
+    return true;
+  }
+
+  const pathIsNotADirectory = !(await stat(projectPath)).isDirectory();
+
+  if (pathIsNotADirectory) {
+    return true;
+  }
+
+  const directoryContent = await readdir(projectPath);
+  const pathIsEmptyDir = directoryContent.length === 0;
+
+  if (pathIsEmptyDir) {
+    return true;
+  }
+
+  const dirHasOnlyGitInside =
+    directoryContent.length === 1 && directoryContent[0] === ".git";
+
+  if (dirHasOnlyGitInside) {
+    return true;
+  }
+
+  const hasUserConfirmedInitInNonEmptyDir = await confirm({
+    message: `Directory ${color.yellow(projectPath)} is not empty. Proceed?`,
+  });
+
+  if (hasUserConfirmedInitInNonEmptyDir) {
+    return true;
+  }
+
+  return false;
 };
 
 type InitTSorJSProjectArg = {