+++ test/integration.test.ts
@@ -139,12 +139,23 @@ describe("integration tests", () => {
         FS_OPTIONS,
       );
 
-      const pathToNewServiceDir = join("src", "services", "newService");
+      const pathToNewServiceDir = join("src", "services", WD_NEW_SERVICE_NAME);
 
       await fluence({
-        args: ["service", "new", "newService"],
+        args: ["service", "new", WD_NEW_SERVICE_NAME],
         cwd,
       });
+
+      const readInterfacesFile = async () => {
+        return readFile(
+          join(cwd, ".fluence", "aqua", "services.aqua"),
+          FS_OPTIONS,
+        );
+      };
+
+      let interfacesFileContent = await readInterfacesFile();
+
+      expect(interfacesFileContent).toBe(`${WD_NEW_SERVICE_INTERFACE}\n`);
 
       const newServiceConfig = await initServiceConfig(
         pathToNewServiceDir,
@@ -155,10 +166,41 @@ describe("integration tests", () => {
       newServiceConfig.modules.facade.envs = { A: "B" };
       await newServiceConfig.$commit();
 
-      const pathToNewSpell = join("src", "spells", "newSpell");
+      const NEW_SPELL_NAME = "newSpell";
+
+      const pathToNewSpell = join("src", "spells", NEW_SPELL_NAME);
+
+      // update first service module source code so it contains a struct
+
+      await writeFile(
+        join(
+          join(cwd, "src", "services", WD_NEW_SERVICE_NAME),
+          join("modules", WD_NEW_SERVICE_NAME, "src", "main.rs"),
+        ),
+        WD_MAIN_RS_CONTENT,
+        FS_OPTIONS,
+      );
 
       await fluence({
-        args: ["spell", "new", "newSpell"],
+        args: ["service", "new", WD_NEW_SERVICE_2_NAME],
+        cwd,
+      });
+
+      interfacesFileContent = await readInterfacesFile();
+
+      expect(interfacesFileContent).toBe(WD_SERVICE_INTERFACES);
+
+      await fluence({
+        args: ["build"],
+        cwd,
+      });
+
+      interfacesFileContent = await readInterfacesFile();
+
+      expect(interfacesFileContent).toBe(WD_UPDATED_SERVICE_INTERFACES);
+
+      await fluence({
+        args: ["spell", "new", NEW_SPELL_NAME],
         cwd,
       });
 
@@ -189,8 +231,11 @@ describe("integration tests", () => {
           fluenceConfig.workers[DEFAULT_WORKER_NAME] !== undefined,
       );
 
-      fluenceConfig.workers[DEFAULT_WORKER_NAME].services = ["newService"];
-      fluenceConfig.workers[DEFAULT_WORKER_NAME].spells = ["newSpell"];
+      fluenceConfig.workers[DEFAULT_WORKER_NAME].services = [
+        WD_NEW_SERVICE_2_NAME,
+      ];
+
+      fluenceConfig.workers[DEFAULT_WORKER_NAME].spells = [NEW_SPELL_NAME];
       await fluenceConfig.$commit();
 
       await fluence({
@@ -283,14 +328,18 @@ describe("integration tests", () => {
       // Jest has a global timeout for each test and if it runs out test will fail
       // eslint-disable-next-line no-constant-condition
       while (true) {
-        const res = await fluence({
-          args: ["run"],
-          flags: {
-            f: "runDeployedServices()",
-            quiet: true,
-          },
-          cwd,
-        });
+        let res = "[]";
+
+        try {
+          res = await fluence({
+            args: ["run"],
+            flags: {
+              f: "runDeployedServices()",
+              quiet: true,
+            },
+            cwd,
+          });
+        } catch {}
 
         const parsedRes = JSON.parse(res);
         assert(Array.isArray(parsedRes));
@@ -341,3 +390,53 @@ const compileAqua = (cwd: string) => {
 const getIndexJSorTSPath = (JSOrTs: "js" | "ts", cwd: string): string => {
   return join(cwd, "src", JSOrTs, "src", `index.${JSOrTs}`);
 };
+
+const WD_MAIN_RS_CONTENT = `#![allow(non_snake_case)]
+use marine_rs_sdk::marine;
+use marine_rs_sdk::module_manifest;
+
+module_manifest!();
+
+pub fn main() {}
+
+#[marine]
+pub struct MyStruct {
+    a: i32,
+    b: i32,
+}
+
+#[marine]
+pub fn greeting() -> MyStruct {
+  MyStruct{
+    a: 1,
+    b: 2,
+  }
+}
+`;
+
+const WD_NEW_SERVICE_NAME = "newService";
+
+const WD_NEW_SERVICE_INTERFACE = `service NewService("${WD_NEW_SERVICE_NAME}"):
+  greeting(name: string) -> string`;
+
+const WD_NEW_SERVICE_2_NAME = "newService2";
+
+const WD_NEW_SERVICE_2_INTERFACE = `service NewService2("${WD_NEW_SERVICE_2_NAME}"):
+  greeting(name: string) -> string`;
+
+const WD_SERVICE_INTERFACES = `${WD_NEW_SERVICE_INTERFACE}
+
+
+${WD_NEW_SERVICE_2_INTERFACE}
+`;
+
+const WD_UPDATED_SERVICE_INTERFACES = `data MyStruct:
+  a: i32
+  b: i32
+
+service NewService("${WD_NEW_SERVICE_NAME}"):
+  greeting() -> MyStruct
+
+
+${WD_NEW_SERVICE_2_INTERFACE}
+`;